+++ src/main/java/io/github/pandier/frostednameplates/Nameplate.java
@@ -9,10 +9,8 @@ import com.github.retrooper.packetevents.wrapper.play.server.WrapperPlayServerEn
 import com.github.retrooper.packetevents.wrapper.play.server.WrapperPlayServerSetPassengers;
 import com.github.retrooper.packetevents.wrapper.play.server.WrapperPlayServerSpawnEntity;
 import io.github.pandier.frostednameplates.util.PacketConsumer;
-import io.github.retrooper.packetevents.adventure.serializer.legacy.LegacyComponentSerializer;
 import io.github.retrooper.packetevents.util.SpigotReflectionUtil;
 import net.kyori.adventure.text.Component;
-import org.bukkit.ChatColor;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 
@@ -22,11 +20,6 @@ import java.util.UUID;
 
 @ApiStatus.Internal
 public class Nameplate {
-    private static final LegacyComponentSerializer LEGACY_SERIALIZER = LegacyComponentSerializer.builder()
-            .character(LegacyComponentSerializer.SECTION_CHAR)
-            .useUnusualXRepeatedCharacterHexFormat()
-            .build();
-
     private static final int FLAGS_DATA_INDEX = 0;
     private static final int CUSTOM_NAME_DATA_INDEX = 2;
     private static final int CUSTOM_NAME_VISIBLE_DATA_INDEX = 3;
@@ -35,14 +28,12 @@ public class Nameplate {
     private final int entityId;
     private final UUID uuid;
     private final int targetEntityId;
-    private String text;
-    private Component textComponent;
+    private Component text = Component.empty();
 
     public Nameplate(int entityId, @NotNull UUID uuid, int targetEntityId) {
         this.entityId = entityId;
         this.uuid = uuid;
         this.targetEntityId = targetEntityId;
-        setText(ChatColor.RED + "N/A");
     }
 
     public Nameplate(int targetEntityId) {
@@ -58,7 +49,7 @@ public class Nameplate {
         packetConsumer.accept(new WrapperPlayServerSpawnEntity(entityId, Optional.of(uuid), EntityTypes.TEXT_DISPLAY, position.add(0.0, 1.8, 0.0), 0f, 0f, 0f, 0, Optional.of(new Vector3d())));
         packetConsumer.accept(new WrapperPlayServerEntityMetadata(entityId, List.of(
                 new EntityData<>(FLAGS_DATA_INDEX, EntityDataTypes.BYTE, getFlags(false)),
-                new EntityData<>(CUSTOM_NAME_DATA_INDEX, EntityDataTypes.OPTIONAL_ADV_COMPONENT, Optional.of(this.textComponent)),
+                new EntityData<>(CUSTOM_NAME_DATA_INDEX, EntityDataTypes.OPTIONAL_ADV_COMPONENT, Optional.of(this.text)),
                 new EntityData<>(CUSTOM_NAME_VISIBLE_DATA_INDEX, EntityDataTypes.BOOLEAN, true),
                 // Just setting the background color will make the text display not display anything, while still showing the custom name
                 new EntityData<>(TEXT_DISPLAY_BACKGROUND_DATA_INDEX, EntityDataTypes.INT, 0)
@@ -68,7 +59,7 @@ public class Nameplate {
 
     public void sendUpdatePackets(@NotNull PacketConsumer packetConsumer) {
         packetConsumer.accept(new WrapperPlayServerEntityMetadata(entityId, List.of(
-                new EntityData<>(CUSTOM_NAME_DATA_INDEX, EntityDataTypes.OPTIONAL_ADV_COMPONENT, Optional.of(this.textComponent))
+                new EntityData<>(CUSTOM_NAME_DATA_INDEX, EntityDataTypes.OPTIONAL_ADV_COMPONENT, Optional.of(this.text))
         )));
     }
 
@@ -83,13 +74,12 @@ public class Nameplate {
         packetConsumer.accept(new WrapperPlayServerDestroyEntities(entityId));
     }
 
-    public void setText(@NotNull String text) {
+    public void setText(@NotNull Component text) {
         this.text = text;
-        this.textComponent = LEGACY_SERIALIZER.deserialize(text);
     }
 
     @NotNull
-    public String getText() {
+    public Component getText() {
         return text;
     }
 }