+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Different formatting options (Legacy, MiniMessage)
+
+### Changed
+
+- Plugin now only supports Paper
+
 ## [0.1.0-alpha.1] - 2025-06-24
 
 Initial release